поработал над AdController

+++ src/main/java/ru/skypro/homework/repository/AdRepository.java
@@ -3,5 +3,5 @@ package ru.skypro.homework.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.skypro.homework.model.AdEntity;
 
-public interface AdRepository extends JpaRepository<AdEntity, Long> {
+public interface AdRepository extends JpaRepository<AdEntity, Integer> {
 }
